Making cw_funcs into a CheckWatertight class.
