+++ tools/make_watertight/cw_func.cpp
@@ -28,10 +28,8 @@
 #include "zip.hpp"
 #include "moab/Skinner.hpp"
 
-namespace cw_func
-{
 
-moab::ErrorCode check_mesh_for_watertightness( moab::EntityHandle input_set, double tol, bool &sealed, bool test, bool verbose, bool check_topology )
+moab::ErrorCode CheckWatertight::check_mesh_for_watertightness( moab::EntityHandle input_set, double tol, bool &sealed, bool test, bool verbose, bool check_topology )
 {
 
   moab::ErrorCode result;
@@ -385,56 +383,55 @@ moab::ErrorCode check_mesh_for_watertightness( moab::EntityHandle input_set, dou
   return moab::MB_SUCCESS;
 }
 
-/* qsort struct comparision function */
-int compare_by_handle(const void *a, const void *b)
-{
-  struct coords_and_id *ia = (struct coords_and_id *)a;
-  struct coords_and_id *ib = (struct coords_and_id *)b;
-  if(ia->vert1 == ib->vert1) {
-    return (int)(ia->vert2 - ib->vert2);
-  } else {
-    return (int)(ia->vert1 - ib->vert1);
-  }
-  /* float comparison: returns negative if b > a
-     and positive if a > b. We multiplied result by 100.0
-     to preserve decimal fraction */
-}
+// /* qsort struct comparision function */
+// int CheckWatertight::compare_by_handle(const void *a, const void *b)
+// {
+//   struct coords_and_id *ia = (struct coords_and_id *)a;
+//   struct coords_and_id *ib = (struct coords_and_id *)b;
+//   if(ia->vert1 == ib->vert1) {
+//     return (int)(ia->vert2 - ib->vert2);
+//   } else {
+//     return (int)(ia->vert1 - ib->vert1);
+//   }
+//   /* float comparison: returns negative if b > a
+//      and positive if a > b. We multiplied result by 100.0
+//      to preserve decimal fraction */
+// }
 
-/* qsort struct comparision function */
-// This is tricky because doubles always get rounded down to ints.
-int compare_by_coords(const void *a, const void *b)
-{
-  struct coords_and_id *ia = (struct coords_and_id *)a;
-  struct coords_and_id *ib = (struct coords_and_id *)b;
-  if(ia->x1 == ib->x1) {
-    if(ia->y1 == ib->y1) {
-      if(ia->z1 == ib->z1) {
-        if(ia->x2 == ib->x2) {
-          if(ia->y2 == ib->y2) {
-            if(ia->z2 == ib->z2) {
-              return ia->surf_id - ib->surf_id;
-            } else {
-              return (ia->z2 > ib->z2) - (ia->z2 < ib->z2);
-            }
-          } else {
-            return (ia->y2 > ib->y2) - (ia->y2 < ib->y2);
-          }
-        } else {
-          return (ia->x2 > ib->x2) - (ia->x2 < ib->x2);
-        }
-      } else {
-        return (ia->z1 > ib->z1) - (ia->z1 < ib->z1);
-      }
-    } else {
-      return (ia->y1 > ib->y1) - (ia->y1 < ib->y1);;
-    }
-  } else {
-    return (ia->x1 > ib->x1) - (ia->x1 < ib->x1);
-  }
-  /* float comparison: returns negative if b > a
-     and positive if a > b. We multiplied result by 100.0
-     to preserve decimal fraction */
-}
+// /* qsort struct comparision function */
+// // This is tricky because doubles always get rounded down to ints.
+// int CheckWatertight::compare_by_coords(const void *a, const void *b)
+// {
+//   struct coords_and_id *ia = (struct coords_and_id *)a;
+//   struct coords_and_id *ib = (struct coords_and_id *)b;
+//   if(ia->x1 == ib->x1) {
+//     if(ia->y1 == ib->y1) {
+//       if(ia->z1 == ib->z1) {
+//         if(ia->x2 == ib->x2) {
+//           if(ia->y2 == ib->y2) {
+//             if(ia->z2 == ib->z2) {
+//               return ia->surf_id - ib->surf_id;
+//             } else {
+//               return (ia->z2 > ib->z2) - (ia->z2 < ib->z2);
+//             }
+//           } else {
+//             return (ia->y2 > ib->y2) - (ia->y2 < ib->y2);
+//           }
+//         } else {
+//           return (ia->x2 > ib->x2) - (ia->x2 < ib->x2);
+//         }
+//       } else {
+//         return (ia->z1 > ib->z1) - (ia->z1 < ib->z1);
+//       }
+//     } else {
+//       return (ia->y1 > ib->y1) - (ia->y1 < ib->y1);;
+//     }
+//   } else {
+//     return (ia->x1 > ib->x1) - (ia->x1 < ib->x1);
+//   }
+//   /* float comparison: returns negative if b > a
+//      and positive if a > b. We multiplied result by 100.0
+//      to preserve decimal fraction */
+// }
 
-}
